Initial Commit
I added a LOT of classes, and have some big plans for this program.
Perhaps everything will work smoothly, perhaps not. Either way, I can't
wait to get to work on this tomorrow, and am forced to commit now so
that I can transfer everything in the morning. Nothing is complete, but
I tried to get in all of the major files and directions in while they
are fresh in my mind.