Added some database input
Third time I've had to commit so that I can take the work on the road...
This should be the last time throughout the weekend though. I forgot to
mention it in the previous commit, but the angle calculation is faulty.
Haven't put any time into fixing it yet, just trying to create the
skeleton of the program before getting to work on the MVP.